+++ modules/mcagar/src/main/ts/ephox/mcagar/api/ThemeSelectors.ts
@@ -1,6 +1,7 @@
 import { Arr, Fun, Global, Obj } from '@ephox/katamari';
 
 import { Editor } from '../alien/EditorTypes';
+import { getSetting } from '../alien/Settings';
 
 const isSilver = (): boolean => {
   const tinymce = Global.tinymce;
@@ -31,7 +32,7 @@ const ModernThemeSelectors: ThemeSelectors = {
 };
 
 const SilverThemeSelectors: ThemeSelectors = {
-  toolBarSelector: (editor: Editor) => Arr.exists([ editor.getParam('toolbar_mode'), editor.getParam('toolbar_drawer') ], (s) => s === 'floating' || s === 'sliding') ? '.tox-toolbar-overlord' : '.tox-toolbar',
+  toolBarSelector: (editor: Editor) => Arr.exists([ getSetting(editor, 'toolbar_mode'), getSetting(editor, 'toolbar_drawer') ], (s) => s === 'floating' || s === 'sliding') ? '.tox-toolbar-overlord' : '.tox-toolbar',
   menuBarSelector: '.tox-menubar',
   dialogSelector: 'div[role="dialog"]',
   dialogCancelSelector: '.tox-button:contains("Cancel")',