+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/toolbar/ToolbarFocusTest.ts
@@ -8,12 +8,12 @@ import { RawEditorOptions } from 'tinymce/core/api/OptionTypes';
 
 describe('browser.tinymce.themes.silver.editor.toolbar.ToolbarFocusTest', () => {
 
-  const pTestFocus = async (settings: RawEditorOptions) => {
+  const pTestFocus = async (options: RawEditorOptions) => {
     const editor = await McEditor.pFromSettings<Editor>({
       toolbar: 'undo redo | bold italic',
       menubar: false,
       statusbar: false,
-      ...settings,
+      ...options,
       base_url: '/project/tinymce/js/tinymce'
     });
     editor.focus();